refactor: convert single file module into a package
