+++ klangbecken/__init__.py
@@ -1,0 +1,3 @@
+from .api import KlangbeckenAPI  # noqa
+
+__version__ = "0.0.12"
