+++ klangbecken/__main__.py
@@ -1,0 +1,3 @@
+from .cli import main
+
+main()
