+++ setup.cfg
@@ -36,7 +36,7 @@ commands = coverage report
 branch = True
 
 [coverage:report]
-include=klangbecken.py
+include=klangbecken/*.py
 ignore_errors = True
 show_missing = True
 # At least 80% of all code must be covered by unit tests
